+++ vm/actor/src/builtin/miner/mod.rs
@@ -6,13 +6,13 @@ mod deadline_assignment;
 mod deadline_state;
 mod deadlines;
 mod expiration_queue;
-mod miner_state;
 mod monies;
 mod partition_state;
 mod policy;
 mod quantize;
 mod sector_map;
 mod sectors;
+mod state;
 mod termination;
 mod types;
 mod vesting_state;
@@ -22,13 +22,13 @@ pub use deadline_assignment::*;
 pub use deadline_state::*;
 pub use deadlines::*;
 pub use expiration_queue::*;
-pub use miner_state::*;
 pub use monies::*;
 pub use partition_state::*;
 pub use policy::*;
 pub use quantize::*;
 pub use sector_map::*;
 pub use sectors::*;
+pub use state::*;
 pub use termination::*;
 pub use types::SectorOnChainInfo;
 pub use types::*;
